add select property interface
