+++ src/notion-api/index.ts
@@ -1,0 +1,11 @@
+import {
+    Color,
+    TextColor,
+    BackgroundColor,
+} from './mixins'
+
+export {
+    Color,
+    TextColor,
+    BackgroundColor,
+}
